Updates to support Python3

+++ env/GapGame1D.py
@@ -681,7 +681,7 @@ class GapGame1D(object):
         # gap_start=random.randint(2,7)
         gap_start=self._terrainParameters['gap_start']
         next_gap=self._terrainParameters['distance_till_next_gap']
-        for i in range(terrainLength/next_gap):
+        for i in range(int(terrainLength/next_gap)):
             gap_start= gap_start+np.random.random_integers(self._terrainParameters['random_gap_start_range'][0],
                                                 self._terrainParameters['random_gap_start_range'][1])
             gap_size=np.random.random_integers(self._terrainParameters['random_gap_width_range'][0],
@@ -727,7 +727,7 @@ class GapGame1D(object):
         # gap_start=random.randint(2,7)
         gap_start=self._terrainParameters['gap_start']
         next_gap=self._terrainParameters['distance_till_next_gap']
-        for i in range(terrainLength/next_gap):
+        for i in range(int(terrainLength/next_gap)):
             gap_start= gap_start+random.randint(self._terrainParameters['random_gap_start_range'][0],
                                                 self._terrainParameters['random_gap_start_range'][1])
             gap_size=random.randint(self._terrainParameters['random_gap_width_range'][0],
